Fix error in getting Auth token for zenhub
The Zenhub auth token retrieval process includes an automated process
for logging in via a github account. The mechanism for this seems to
have changed so that the process needed adaption.

The change includes the order of the login process. After this change
the process starts at the zenhub page and tries to login via a github
token.

Furthermore in case of an error we log the current page as downloadable
artefact.